Renaming the CourseScanComponent.tsx to LastScanInfo.tsx
