+++ frontend/app/components/LastScanInfo.tsx
